The Index class is being phased out as it has no more functionality than an STL vector. In this update, it is reduced to a subclass of vector<int>.
Additionally some Makefiles in the bugs module have been modified to
remove the redundant addition of src/modules/bugs/rmath to the
include path.

+++ src/include/compiler/Compiler.h
@@ -30,7 +30,7 @@ class Compiler {
   CounterTab _countertab;
   std::map<std::string, SArray> const &_data_table;
   std::map<std::string, std::vector<bool> > _constant_mask;
-  unsigned long _n_resolved, _n_relations;
+  unsigned int _n_resolved, _n_relations;
   bool *_is_resolved;
   bool _strict_resolution;
   ConstantFactory _constantfactory;
@@ -69,7 +69,7 @@ class Compiler {
   void addDevianceNode();
 public:
   Graph &graph() const;
-  bool indexExpression(ParseTree const *t, long &value);
+  bool indexExpression(ParseTree const *t, int &value);
   MixtureFactory &mixtureFactory();
   SymTab &symTab() const;
   Node* getParameter(ParseTree const *t);
